大致完成！

+++ app/src/main/java/com/example/logisticsystem/ui/query/QueryFragment.kt
@@ -1,24 +1,32 @@
 package com.example.logisticsystem.ui.query
 
 import android.annotation.SuppressLint
+import android.app.AlertDialog
 import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.example.logisticsystem.LogisticView
 import com.example.logisticsystem.MyDatabaseHelper
 import com.example.logisticsystem.databinding.FragmentQueryBinding
 import com.example.logisticsystem.ui.SharedViewModel
+import java.lang.Thread.sleep
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
 
 class QueryFragment : Fragment() {
 
     private var _binding: FragmentQueryBinding? = null
     private val binding get() = _binding!!
 
+    @RequiresApi(Build.VERSION_CODES.O)
     @SuppressLint("LongLogTag")
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -34,7 +42,13 @@ class QueryFragment : Fragment() {
 
         //数据库
         val dbHelper =
-            getActivity()?.let { MyDatabaseHelper(it.getApplicationContext(), "LogisticSystem.db", 1) }
+            getActivity()?.let {
+                MyDatabaseHelper(
+                    it.getApplicationContext(),
+                    "LogisticSystem.db",
+                    1
+                )
+            }
 
         var db = dbHelper?.writableDatabase
 
@@ -65,25 +79,82 @@ class QueryFragment : Fragment() {
         /**
          * 本地运单
          */
-        binding.local.setOnClickListener{
-            val intent=Intent(activity, LogisticView::class.java)
+        binding.local.setOnClickListener {
+            val intent = Intent(activity, LogisticView::class.java)
+            intent.putExtra("mode", "local")
             startActivity(intent)
         }
         /**
+         * xml运单
+         */
+        Log.e("visibility", "visibility改变成INVISIBLE")
+
+
+        binding.remoteXml.setOnClickListener {
+
+            //让progressBar可见
+            binding.progress.visibility = View.VISIBLE
+
+            val intent = Intent(activity, LogisticView::class.java)
+            intent.putExtra("mode", "xml")
+            startActivity(intent)
+        }
+        /**
+         * json运单
+         */
+        binding.remoteJson.setOnClickListener {
+            //让progressBar可见
+            binding.progress.visibility = View.VISIBLE
+            val intent = Intent(activity, LogisticView::class.java)
+            intent.putExtra("mode", "json")
+            startActivity(intent)
+        }
+
+
+        /**
          * 清空本地运单
          */
-        binding.delete.setOnClickListener{
-            if (dbHelper != null) {
-                if (db != null) {
-                    dbHelper.delete(db)
+        binding.delete.setOnClickListener {
+            AlertDialog.Builder(getActivity()).apply {
+                setTitle("通知")
+                setMessage("您确认要清空本地运单吗？")
+                setCancelable(false)
+                //点击back无法返回，默认是true
+                setPositiveButton("确认") { dialog, which ->
+                    if (dbHelper != null) {
+                        if (db != null) {
+                            dbHelper.delete(db)
+                        }
+                    }
+                    Toast.makeText(activity, "清空成功", Toast.LENGTH_SHORT).show()
                 }
+                setNegativeButton("返回") { dialog, which ->
+
+                }
+                show()
             }
+
         }
         return root
     }
 
+    /**
+     * onStart时让progressbar不可见
+     */
+    override fun onStart() {
+        super.onStart()
+        binding.progress.visibility = View.INVISIBLE
+        Log.e("queryFragment被onStart","queryFragment被onStart")
+
+    }
+
+    override fun onPause() {
+        super.onPause()
+    }
+
     override fun onDestroyView() {
         super.onDestroyView()
+
         _binding = null
     }
 }
